+++ zebra-consensus/src/error.rs
@@ -162,6 +162,9 @@ pub enum TransactionError {
     #[error("adding to the sprout pool is disabled after Canopy")]
     DisabledAddToSproutPool,
 
+    #[error("adding to the orchard pool is disabled after NU6.3")]
+    DisabledAddToOrchardPool,
+
     #[error("could not calculate the transaction fee")]
     IncorrectFee,
 
@@ -358,6 +361,7 @@ impl TransactionError {
             | RedPallas(_)
             | BothVPubsNonZero
             | DisabledAddToSproutPool
+            | DisabledAddToOrchardPool
             | NotEnoughFlags
             | WrongConsensusBranchId
             | MissingConsensusBranchId